[ObjC] Split ImportWriter & TextFormatDecodeData into files.
