+++ src/google/protobuf/compiler/objectivec/enum.cc
@@ -38,6 +38,7 @@
 #include "absl/strings/str_cat.h"
 #include "google/protobuf/compiler/objectivec/names.h"
 #include "google/protobuf/compiler/objectivec/helpers.h"
+#include "google/protobuf/compiler/objectivec/text_format_decode_data.h"
 #include "google/protobuf/io/printer.h"
 
 namespace google {
